refactor(product-cdc): update WorkflowType and OnboardingStatus to use SDK
