+++ apps/product-cdc/pom.xml
@@ -28,7 +28,7 @@
     <quarkus-openapi-generator.version>2.16.0</quarkus-openapi-generator.version>
     <quarkus-cucumber-version>1.3.0</quarkus-cucumber-version>
     <selfcare-cucumber-sdk.version>0.1.1</selfcare-cucumber-sdk.version>
-    <selfcare-onboarding-sdk.version>0.15.5</selfcare-onboarding-sdk.version>
+    <selfcare-onboarding-sdk.version>0.15.6</selfcare-onboarding-sdk.version>
     <mapstruct.lombok.version>0.2.0</mapstruct.lombok.version>
   </properties>
   <dependencyManagement>
