+++ apps/product-cdc/src/main/java/it/pagopa/selfcare/product/model/EmailTemplate.java
@@ -1,8 +1,8 @@
 package it.pagopa.selfcare.product.model;
 
 import it.pagopa.selfcare.product.model.enums.InstitutionType;
-import it.pagopa.selfcare.product.model.enums.OnboardingStatus;
-import it.pagopa.selfcare.product.model.enums.WorkflowType;
+import it.pagopa.selfcare.onboarding.common.OnboardingStatus;
+import it.pagopa.selfcare.onboarding.common.WorkflowType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
