--- apps/product-cdc/src/main/java/it/pagopa/selfcare/product/model/enums/OnboardingStatus.java
@@ -1,11 +1,0 @@
-package it.pagopa.selfcare.product.model.enums;
-
-public enum OnboardingStatus {
-  REQUEST,
-  TOBEVALIDATED,
-  PENDING,
-  COMPLETED,
-  FAILED,
-  REJECTED,
-  DELETED
-}
