--- apps/product-cdc/src/main/java/it/pagopa/selfcare/product/model/enums/WorkflowType.java
@@ -1,19 +1,0 @@
-package it.pagopa.selfcare.product.model.enums;
-
-public enum WorkflowType {
-  CONTRACT_REGISTRATION, // INSTITUTION
-  FOR_APPROVE, // INSTITUTION
-  FOR_APPROVE_PT, // INSTITUTION
-  FOR_APPROVE_GPU, // INSTITUTION
-  CONFIRMATION, // INSTITUTION
-  USERS, // USER
-  USERS_IMPORT, // USER
-  IMPORT, // INSTITUTION
-  IMPORT_AGGREGATION,
-  CONTRACT_REGISTRATION_AGGREGATOR, // AGGREGATION
-  CONFIRMATION_AGGREGATE, // AGGREGATION
-  INCREMENT_REGISTRATION_AGGREGATOR, // AGGREGATION
-  CONFIRMATION_AGGREGATOR, // AGGREGATION
-  USERS_PG, // USER
-  USERS_EA // USER
-}
